% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{fit_commonfactor_gwas_dwls}
\alias{fit_commonfactor_gwas_dwls}
\title{Fit the marker-scaled common-factor GWAS DWLS slice used by
GenomicSEM's \code{commonfactorGWAS()} model.}
\usage{
fit_commonfactor_gwas_dwls(sample_cov, wls_v, k, max_iter, tol)
}
\arguments{
\item{sample_cov}{Flattened observed covariance matrix.}

\item{wls_v}{Flattened DWLS weight matrix.}

\item{k}{Number of trait indicators.}

\item{max_iter}{Maximum optimizer iterations.}

\item{tol}{Convergence tolerance.}
}
\description{
Variable order is expected to be \verb{SNP, trait_1, ..., trait_k}.
}
