% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{fitted_rust}
\alias{fitted_rust}
\title{Rust-backed experimental replacement for \code{stats::fitted()}.}
\usage{
fitted_rust(object, ...)
}
\arguments{
\item{object}{A \code{lavaan_rust_fit} object.}

\item{...}{Reserved for lavaan-compatible arguments.}
}
\description{
Rust-backed experimental replacement for \code{stats::fitted()}.
}
