% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{lav_model_get_parameters_rust}
\alias{lav_model_get_parameters_rust}
\title{Rust-backed experimental replacement for lavaan's parameter extractor.}
\usage{
lav_model_get_parameters_rust(...)
}
\arguments{
\item{...}{Reserved for a future compatible implementation.}
}
\description{
Rust-backed experimental replacement for lavaan's parameter extractor.
}
