% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{lavaan_rust_fit-class}
\alias{lavaan_rust_fit-class}
\title{Minimal compatibility fit object for the supported lavaanrust surface.}
\description{
Minimal compatibility fit object for the supported lavaanrust surface.
}
