% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{lavInspect_rust}
\alias{lavInspect_rust}
\title{Rust-backed experimental replacement for \code{lavaan::lavInspect()}.}
\usage{
lavInspect_rust(object, what, ...)
}
\arguments{
\item{object}{A \code{lavaan_rust_fit} object.}

\item{what}{Inspection key.}

\item{...}{Reserved for lavaan-compatible arguments.}
}
\description{
Rust-backed experimental replacement for \code{lavaan::lavInspect()}.
}
