% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{parTable_rust}
\alias{parTable_rust}
\title{Rust-backed experimental replacement for \code{lavaan::parTable()}.}
\usage{
parTable_rust(object, ...)
}
\arguments{
\item{object}{A \code{lavaan_rust_fit} object.}

\item{...}{Reserved for lavaan-compatible arguments.}
}
\description{
Rust-backed experimental replacement for \code{lavaan::parTable()}.
}
